7.2.Обзор брокеров сообщений
