+++ thinktetika/main/urls.py
@@ -21,10 +21,13 @@ from django.conf.urls import include
 from django.urls import path
 
 from .validators.validators import group_check
-from .views import index, GoodsListView, GoodsDetalView, CreateProduct, UpdateProduct
+# from .views import index, GoodsListView, GoodsDetalView, CreateProduct, UpdateProduct
+from .views import GoodsListView, GoodsDetalView, CreateProduct, UpdateProduct, IndexView
 from main.models import Product
 
+
 urlpatterns = [
+                  path('', IndexView.as_view(), name='index'),
                   path('goods/', GoodsListView.as_view(model=Product), name='goods'),
                   path('goods/<int:pk>/', GoodsDetalView.as_view(model=Product), name='good-detail'),
                   path('goods/add', group_check(CreateProduct.as_view()), name='good-add'),
